python/prompt-management/requirement.txt
maxim-py==1.5.13
requests